Remove old settings screen and files
