--- css/instant-articles-settings-wizard.css
@@ -1,3 +1,0 @@
-.instant-articles-inline-form input[name='submit'] {
-	display: inline;
-}
