+++ facebook-instant-articles.php
@@ -74,7 +74,6 @@ if ( version_compare( PHP_VERSION, '5.4', '<' ) ) {
 
 	require_once( dirname( __FILE__ ) . '/embeds.php' );
 	require_once( dirname( __FILE__ ) . '/class-instant-articles-post.php' );
-	require_once( dirname( __FILE__ ) . '/settings/class-instant-articles-settings.php' );
 	require_once( dirname( __FILE__ ) . '/wizard/class-instant-articles-wizard.php' );
 	require_once( dirname( __FILE__ ) . '/meta-box/class-instant-articles-meta-box.php' );
 	require_once( dirname( __FILE__ ) . '/class-instant-articles-publisher.php' );
@@ -89,6 +88,23 @@ if ( version_compare( PHP_VERSION, '5.4', '<' ) ) {
 		flush_rewrite_rules();
 	}
 	register_activation_hook( __FILE__, 'instant_articles_activate' );
+
+	add_action( 'admin_init', 'instant_articles_redirect_settings' );
+
+	/**
+	 * Show a message to set up the plugin when it is activated
+	 */
+	add_action( 'admin_notices', 'instant_articles_setup_admin_notice' );
+	add_action( 'network_admin_notices', 'instant_articles_setup_admin_notice' ); // also show message on multisite
+	function instant_articles_setup_admin_notice() {
+		global $pagenow;
+		if ( $pagenow === 'plugins.php' && Instant_Articles_Wizard_State::get_current_state() !== Instant_Articles_Wizard_State::STATE_REVIEW_SUBMISSION ) {
+			$settings_url = Instant_Articles_Wizard::get_url();
+			echo '<div id="error" class="error success notice is-dismissible">';
+			echo '<p>Your Instant Articles setup is not complete. Go to the <a href="' . esc_url_raw($settings_url) . '">Instant Articles Settings</a> page to complete it.';
+			echo '</div>';
+		}
+	}
 
 	/**
 	 * Plugin activation hook to remove our rewrite rules.
@@ -188,7 +204,6 @@ if ( version_compare( PHP_VERSION, '5.4', '<' ) ) {
 				) );
 			}
 		}
-
 	}
 	add_action( 'pre_get_posts', 'instant_articles_query', 10, 1 );
 
@@ -226,10 +241,6 @@ if ( version_compare( PHP_VERSION, '5.4', '<' ) ) {
 		wp_register_style(
 			'instant-articles-meta-box',
 			plugins_url( '/css/instant-articles-meta-box.css', __FILE__ )
-		);
-		wp_register_style(
-			'instant-articles-settings-wizard',
-			plugins_url( '/css/instant-articles-settings-wizard.css', __FILE__ )
 		);
 		wp_register_style(
 			'instant-articles-settings',
@@ -319,9 +330,6 @@ if ( version_compare( PHP_VERSION, '5.4', '<' ) ) {
 		}
 	}
 	add_action( 'wp_head', 'inject_url_claiming_meta_tag' );
-
-	// Initialize the Instant Articles settings page.
-	Instant_Articles_Settings::init();
 
 	// Initialize the Instant Articles meta box.
 	Instant_Articles_Meta_Box::init();
