+++ wizard/class-instant-articles-option-publishing.php
